requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers & Tokenizers
transformers>=4.44.0
tokenizers>=0.19.0

# PEFT (Parameter-Efficient Fine-Tuning)
peft>=0.12.0

# Quantization
bitsandbytes>=0.43.0

# Dataset Handling
datasets>=2.18.0
accelerate>=0.30.0

# HuggingFace Hub
huggingface-hub>=0.23.0

# Model Evaluation
scikit-learn>=1.4.0
nltk>=3.8.1

# Configuration & Utils
PyYAML>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Development & Testing (optional)
pytest>=8.0.0
jupyter>=1.0.0
tensorboard>=2.15.0

# Additional Tools
sentencepiece>=0.1.99
protobuf>=4.21.0
